CP-54291 Move refcounter, journaler and fjournaler from /opt
