+++ libs/sm/cleanup.py
@@ -37,11 +37,11 @@ from sm import lvutil
 from sm import vhdutil
 from sm import lvhdutil
 import lvmcache
-import journaler
-import fjournaler
+from sm import journaler
+from sm import fjournaler
 from sm.core import lock
 from sm.core import util
-from refcounter import RefCounter
+from sm.refcounter import RefCounter
 from lvmanager import LVActivator
 from srmetadata import LVMMetadataHandler
